pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vhh-paratope-clustering"
version = "0.1.0"
description = "A tool for structural paratope clustering of VHH (nanobody) antibodies"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "ANARCI",
    "ImmuneBuilder",
    "biopython>=1.79",
    "numpy>=1.21",
    "scipy>=1.7",
    "scikit-learn>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["vhh_paratope_clustering*"]
